+++ app/assets/javascripts/folio/console/modules/with_aside.js
@@ -1,0 +1,6 @@
+window.Folio.Stimulus.register('f-c-with-aside', class extends window.Stimulus.Controller {
+  toggle (e) {
+    e.preventDefault()
+    this.element.classList.toggle('f-c-with-aside--aside-collapsed')
+  }
+})