+++ app/cells/folio/console/catalogue_cell.rb
@@ -167,8 +167,9 @@ class Folio::Console::CatalogueCell < Folio::ConsoleCell
   end
 
   def toggle(attr, opts = {})
-    attribute(attr, class_name: "toggle") do
-      cell("folio/console/boolean_toggle", record, opts.merge(attribute: attr))
+    attribute(attr, class_name: "toggle", aligned: true) do
+      render_view_component(Folio::Console::Ui::BooleanToggleComponent.new(record:,
+                                                                           attribute: attr))
     end
   end