+++ app/controllers/concerns/folio/console/default_actions.rb
@@ -284,7 +284,7 @@ module Folio::Console::DefaultActions
     end
 
     def response_with_json_for_valid_update
-      if params[:_trigger] == "f-c-boolean-toggle"
+      if params[:_trigger] == "f-c-ui-boolean-toggle"
         render json: {
           data: {
             f_c_catalogue_published_dates: cell("folio/console/catalogue/published_dates", folio_console_record).show,