+++ app/models/folio/atom/base.rb
@@ -54,15 +54,21 @@ class Folio::Atom::Base < Folio::ApplicationRecord
   validate :validate_placement
 
   def self.cell_name
-    nil
+  end
+
+  def self.component_class
+    if cell_name.nil?
+      "#{self.class}Component".constantize
+    end
+  end
+
+  def self.molecule_component_class
   end
 
   def self.splittable_by_attribute
-    nil
   end
 
   def cell_options
-    nil
   end
 
   def partial_name