+++ lib/generators/folio/molecule/molecule_generator.rb
@@ -1,10 +1,8 @@
 # frozen_string_literal: true
 
-require Folio::Engine.root.join("lib/generators/folio/generator_base")
+require Folio::Engine.root.join("lib/generators/folio/atom/atom_generator")
 
-class Folio::MoleculeGenerator < Rails::Generators::NamedBase
-  include Folio::GeneratorBase
-
+class Folio::MoleculeGenerator < Folio::AtomGenerator
   source_root File.expand_path("templates", __dir__)
 
   def atom_model
@@ -12,12 +10,26 @@ class Folio::MoleculeGenerator < Rails::Generators::NamedBase
   end
 
   def cell
-    template "cell.rb.tt", "app/cells/#{application_namespace_path}/molecule/#{name}_cell.rb"
-    template "cell.slim.tt", "app/cells/#{application_namespace_path}/molecule/#{name}/show.slim"
-    template "cell_test.rb.tt", "test/cells/#{application_namespace_path}/molecule/#{name}_cell_test.rb"
+    if options[:cell]
+      template "cell.rb.tt", "app/cells/#{application_namespace_path}/molecule/#{name}_cell.rb"
+      template "cell.slim.tt", "app/cells/#{application_namespace_path}/molecule/#{name}/show.slim"
+      template "cell_test.rb.tt", "test/cells/#{application_namespace_path}/molecule/#{name}_cell_test.rb"
+    end
   end
 
-  def i18n_yml
-    add_atom_to_i18n_ymls
+  def component
+    unless options[:cell]
+      template "component.rb.tt", "app/components/#{application_namespace_path}/molecule/#{name}_component.rb"
+      template "component.slim.tt", "app/components/#{application_namespace_path}/molecule/#{name}_component.slim"
+      template "component_test.rb.tt", "test/components/#{application_namespace_path}/molecule/#{name}_component_test.rb"
+    end
+  end
+
+  def molecule_css_class_name
+    "#{classname_prefix}-molecule-#{dashed_resource_name}"
+  end
+
+  def molecule_component_name
+    "#{application_namespace}::Molecule::#{class_name}Component"
   end
 end