+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 2023-10-02
+### Changed
+- atom and molecule generator now generates components by default - to use cells, pass the `-cell` option
+- renamed `cell_options` to `atom_options` in `render_atoms` and `render_atoms_in_molecules`
+
+## 2023-09-27
+### Added
+- added `render_view_component` to `Folio::ApplicationCell`
+- added `Folio::Console::Ui::BooleanToggleComponent`
+### Removed
+- removed `Folio::Console::BooleanToggleCell` - replace your `folio/console/boolean_toggle` cell calls with `Folio::Console::Ui::BooleanToggleComponent`
+
 ## 2023-07-28
 ### Changed
 - version 4.0!